+++ lib/rspec/core/configuration.rb
@@ -1,7 +1,7 @@
 require 'fileutils'
 require 'rspec/core/backtrace_formatter'
 require 'rspec/core/ruby_project'
-require 'rspec/core/formatters/deprecation_formatter.rb'
+require 'rspec/core/formatters/deprecation_formatter'
 
 module RSpec
   module Core
@@ -1009,6 +1009,20 @@ module RSpec
       end
 
       module ExposeCurrentExample; end
+
+      # Turns deprecation warnings into errors, in order to surface
+      # the full backtrace of the call site. This can be useful when
+      # you need more context to address a deprecation than the
+      # single-line call site normally provided.
+      #
+      # @example
+      #
+      #   RSpec.configure do |rspec|
+      #     rspec.raise_errors_for_deprecations!
+      #   end
+      def raise_errors_for_deprecations!
+        self.deprecation_stream = Formatters::DeprecationFormatter::RaiseErrorStream.new
+      end
 
     private